Fix load/store US/CS when eew!=sew
Previously we misunderstood the vector spec, using eew&lmul to control
load/store US/CS. Per section 7.3 in the spec, we should instead use
eew&emul to control it (as if performing a mixed-precision operation
with a fictional operand).

After reverting the incorrect RVV side change, this change modifies
the LSU to operate on correct emul instead of the lmul.

Affected tests are re-enabled.

Change-Id: If1298882bb4acf65fc847fc6153a3dc983e5f813