Move manual tag from single tests to test suites
Previously we discovered that verilator tests don't benefit much from
multithreading. It makes more sense to run multiple tests in parallel.
Because the test suite target doesn't do this, we need to do it through
bazel.

This change marks the verilator and vcs test suites as manual, and
single tests as not manual.

Change-Id: I435234c5e012cb5c03cc05df1a6a70ce8bacbe3c